create controller

+++ src/main/java/amcom/desafio/orderms/entity/OrderItem.java
@@ -3,18 +3,20 @@ package amcom.desafio.orderms.entity;
 import org.springframework.data.mongodb.core.mapping.Field;
 import org.springframework.data.mongodb.core.mapping.FieldType;
 
+import java.math.BigDecimal;
+
 public class OrderItem {
 
     private String product;
     private Integer quantity;
 
     @Field(targetType = FieldType.DECIMAL128)
-    private Double price;
+    private BigDecimal price;
 
     public OrderItem() {
     }
 
-    public OrderItem(String product, Integer quantity, Double price) {
+    public OrderItem(String product, Integer quantity, BigDecimal price) {
         this.product = product;
         this.quantity = quantity;
         this.price = price;
@@ -36,11 +38,11 @@ public class OrderItem {
         this.quantity = quantity;
     }
 
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
 }
